Switch kubeconfig-secret webhook
