+++ pkg/admissioncontroller/webhook/add.go
@@ -50,6 +50,12 @@ func AddToManager(
 		return err
 	}
 
+	if err := (&kubeconfigsecret.Handler{
+		Logger: mgr.GetLogger().WithName("webhook").WithName(kubeconfigsecret.HandlerName),
+	}).AddToManager(mgr); err != nil {
+		return fmt.Errorf("failed adding %s webhook handler: %w", kubeconfigsecret.HandlerName, err)
+	}
+
 	if err := (&namespacedeletion.Handler{
 		Logger:    mgr.GetLogger().WithName("webhook").WithName(namespacedeletion.HandlerName),
 		APIReader: mgr.GetAPIReader(),
@@ -72,7 +78,6 @@ func AddToManager(
 
 	server.Register(seedauthorizer.WebhookPath, seedauthorizer.NewHandler(logSeedAuth, seedauthorizer.NewAuthorizer(logSeedAuth, graph)))
 	server.Register(seedrestriction.WebhookPath, &webhook.Admission{Handler: seedRestrictionHandler, RecoverPanic: true})
-	server.Register(kubeconfigsecret.WebhookPath, &webhook.Admission{Handler: kubeconfigsecret.New(log.WithName(kubeconfigsecret.HandlerName)), RecoverPanic: true})
 	server.Register(auditpolicy.WebhookPath, &webhook.Admission{Handler: auditpolicy.New(log.WithName(auditpolicy.HandlerName)), RecoverPanic: true})
 	server.Register(internaldomainsecret.WebhookPath, &webhook.Admission{Handler: internaldomainsecret.New(log.WithName(internaldomainsecret.HandlerName)), RecoverPanic: true})
