--- pkg/logger/logger.go
@@ -1,32 +1,0 @@
-// Copyright (c) 2021 SAP SE or an SAP affiliate company. All rights reserved. This file is licensed under the Apache Software License, v. 2 except as noted otherwise in the LICENSE file
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package logger
-
-import (
-	"github.com/go-logr/logr"
-	"k8s.io/apimachinery/pkg/util/runtime"
-
-	"github.com/gardener/gardener/pkg/utils"
-)
-
-// IDFieldName is the name of the id field for a logger.
-const IDFieldName = "process_id"
-
-// NewIDLogger extends an existing logger with a randomly generated id field (key `process_id`).
-func NewIDLogger(logger logr.Logger) logr.Logger {
-	id, err := utils.GenerateRandomString(8)
-	runtime.Must(err)
-	return logger.WithValues(IDFieldName, id)
-}
