--- pkg/logger/logger_test.go
@@ -1,34 +1,0 @@
-// Copyright (c) 2021 SAP SE or an SAP affiliate company. All rights reserved. This file is licensed under the Apache Software License, v. 2 except as noted otherwise in the LICENSE file
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package logger_test
-
-import (
-	. "github.com/gardener/gardener/pkg/logger"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-	"github.com/onsi/gomega/gbytes"
-	logzap "sigs.k8s.io/controller-runtime/pkg/log/zap"
-)
-
-var _ = Describe("#NewIDLogger", func() {
-	It("should return an Entry object an ID field", func() {
-		logBuffer := gbytes.NewBuffer()
-		logger := logzap.New(logzap.JSONEncoder(), logzap.WriteTo(logBuffer))
-
-		NewIDLogger(logger).Info("Foo")
-		Eventually(logBuffer).Should(gbytes.Say(`"msg":"Foo","process_id":"[a-zA-Z0-9]{8}"`))
-	})
-})
