+++ pkg/logger/zap.go
@@ -66,25 +66,3 @@ func NewZapLogger(level string, format string, additionalOpts ...logzap.Opts) (l
 
 	return logzap.New(append(opts, additionalOpts...)...), nil
 }
-
-// ZapLogger is a Logger implementation.
-// If development is true, a Zap development config will be used
-// (stacktraces on warnings, no sampling), otherwise a Zap production
-// config will be used (stacktraces on errors, sampling).
-// Additionally, the time encoding is adjusted to `zapcore.ISO8601TimeEncoder`.
-// This is used by extensions for historical reasons.
-// TODO: consolidate this with NewZapLogger and make everything configurable in a harmonized way
-func ZapLogger(development bool) logr.Logger {
-	return logzap.New(func(o *logzap.Options) {
-		var encCfg zapcore.EncoderConfig
-		if development {
-			encCfg = zap.NewDevelopmentEncoderConfig()
-		} else {
-			encCfg = zap.NewProductionEncoderConfig()
-		}
-		setCommonEncoderConfigOptions(&encCfg)
-
-		o.Encoder = zapcore.NewJSONEncoder(encCfg)
-		o.Development = development
-	})
-}
